mujoco/requirements.txt
bleach==1.5.0
getch==1.0
html5lib==0.9999999
keyboard==0.13.2
Markdown==2.6.11
numpy==1.14.5
pkg-resources==0.0.0
protobuf==3.6.0
six==1.11.0
tensorflow_gpu==1.3
tensorflow-tensorboard==0.1.8
Werkzeug==0.14.1
tqdm
gym
mujoco_py
opencv-python==3.4.1.15
